fix state space type eror

+++ sdzoo/env/sdzoo.py
@@ -232,9 +232,9 @@ class parallel_env(ParallelEnv):
         if observe_method in ["pyg"]:
             # deficit, surplus
             state_space["current_node"] = spaces.Box(
-                low = np.array([-1, -1], dtype=np.int32),
-                high = np.array([np.inf, np.inf], dtype=np.int32),
-                dtype=np.int32
+                low = np.array([-1.0, -1.0], dtype=np.float32),
+                high = np.array([np.inf, np.inf], dtype=np.float32),
+                dtype=np.float32
             )
 
         if observe_method in ["pyg"]:
